refactor: moved warnings into the util folder
