+++ src/plugin_system/plugin_system.cpp
@@ -1,4 +1,4 @@
-#include <warnings.h>
+#include <util/warnings.h>
 #include "plugin_system.h"
 
 namespace aaltitoad::plugins {
